Make all tabSize properties use NSUInteger as their type

+++ Frameworks/OakTextView/src/OTVStatusBar.h
@@ -13,7 +13,7 @@
 @property (nonatomic) NSString* fileType; // This will update grammarName
 @property (nonatomic, getter = isRecordingMacro) BOOL recordingMacro;
 @property (nonatomic) BOOL softTabs;
-@property (nonatomic) int32_t tabSize;
+@property (nonatomic) NSUInteger tabSize;
 
 @property (nonatomic, weak) id <OTVStatusBarDelegate> delegate;
 @property (nonatomic, weak) id target;
